refactor: BoardInterface를 board 패키지로 위치 변경
