+++ src/main/java/janggi/domain/piece/Piece.java
@@ -1,6 +1,6 @@
 package janggi.domain.piece;
 
-import janggi.domain.BoardInterface;
+import janggi.domain.board.BoardInterface;
 import janggi.domain.Position;
 
 import java.util.List;
